+++ Minimal/Code/Enums/Entities.py
@@ -6,10 +6,6 @@ class Entity:
 
 # --------------------------------------------------
 class Entities(_Enum):
-	@staticmethod
-	def _generate_next_value_(name: string, start: _int, count: _int, last_values: _list[AnyTFWR]) -> Entity:
-		...
-
 	Apple: Entity
 	# Docstring: Apple (Entities)