+++ Minimal/Code/Enums/Hats.py
@@ -6,10 +6,6 @@ class Hat:
 
 # --------------------------------------------------
 class Hats(_Enum):
-	@staticmethod
-	def _generate_next_value_(name: string, start: _int, count: _int, last_values: _list[AnyTFWR]) -> Hat:
-		...
-
 	Brown_Hat: Hat
 	# Docstring: Brown_Hat