+++ Minimal/Code/Enums/Items.py
@@ -6,10 +6,6 @@ class Item:
 
 # --------------------------------------------------
 class Items(_Enum):
-    @staticmethod
-    def _generate_next_value_(name: string, start: _int, count: _int, last_values: _list[AnyTFWR]) -> Item:
-        ...
-
     Bone: Item
     # Docstring: Bone (Items)