+++ Minimal/Code/Enums/Leaderboard.py
@@ -6,10 +6,6 @@ class Leaderboard:
 
 # --------------------------------------------------
 class Leaderboards(_Enum):
-	@staticmethod
-	def _generate_next_value_(name: string, start: _int, count: _int, last_values: _list[AnyTFWR]) -> Leaderboard:
-		...
-
 	Cactus: Leaderboard
 	# Docstring: Cactus (Leaderboard)