+++ Template/Imports.py
@@ -1,12 +1,22 @@
 # Expose some useful types to allow for typing without using a typing import.
 # Typing imports would fail to run in-game as they are not ignored.
+# Ignore any type warnings as needed by your environment.
+# Some types are provided as private to the let file have control
+# over the regular defaults and are not commonly useful for typing most code
 
 # Notes on aliases because of TFWR functions:
 # - string -> builtins.str
 # - range_class -> builtins.range
 
-from typing import Literal, Final, overload
-from collections.abc import Callable, Iterator, Iterable, Sequence, Container
+from typing import (
+    Any, Literal, Final,
+    overload, Self, Never
+)
+from collections.abc import (
+    Callable, Iterator, Iterable,
+    Sequence, Container, Collection
+)
+from types import ModuleType
 
 from builtins import (
     bool, int, float, str as string,
@@ -24,7 +34,4 @@ from builtins import (
     bool as _bool, int as _int, float as _float, range as _range,
     tuple as _tuple, list as _list, set as _set, dict as _dict
 )
-
-from types import ModuleType
-from typing import Any, Self as _Self, Never as _Never
 from enum import Enum as _Enum